+++ MarkdownMonster/Windows/Helpers/ValueConverters.cs
@@ -110,6 +110,42 @@ namespace MarkdownMonster.Windows
         }
     }
 
+    [ValueConversion(typeof(string), typeof(bool))]
+    public class NotEmptyStringToBooleanConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            string text = value as string;
+            if (!string.IsNullOrEmpty(text))
+                return true;
+
+            return false;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+
+    [ValueConversion(typeof(string), typeof(bool))]
+    public class EmptyStringToBooleanConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            string text = value as string;
+            if (string.IsNullOrEmpty(text))
+                return true;
+
+            return false;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+
     public class FileNameFromPathConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)